taller-de-ingestion-de-api-con-dlt/pipelines/openlibrary-pipeline/pyproject.toml
[project]
name = "openlibrary-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "altair>=6.0.0",
    "dlt[workspace]>=1.22.0",
    "ibis-framework[duckdb]>=12.0.0",
    "marimo>=0.20.1",
]
